renames dimcat.data => dimcat.resources and moves features to dimcat.resources.features
