--- src/dimcat/features/__init__.py
@@ -1,3 +1,0 @@
-from .annotations import Annotations, KeyAnnotations
-from .base import Feature, FeatureName
-from .notes import Notes
