--- src/dimcat/features/annotations.py
@@ -1,9 +1,0 @@
-from dimcat.features.base import Feature
-
-
-class Annotations(Feature):
-    pass
-
-
-class KeyAnnotations(Annotations):
-    pass
