+++ src/dimcat/resources/__init__.py
@@ -1,0 +1,1 @@
+from .features import FeatureName
